Restored the anonymous ID code
It looks like UWP apps don't get any sort of ANID-equivalent (or I can't find it) so I'm using whatever this "DeviceFrom" property is (it's "Unknown" on PC and the emulator, so maybe I shouldn't use it?). Regardless, that part is stored away in the Utilities method so I can update it later easily.