Reimplemented most of the phone home code
The "there's a new build available" notification is commented out, but it shouldn't really be necessary anyway since the store ought to manage that for us. I would need to implement a DisplayYesNoBox-type method in a way that's accessible with no context as well as a method to open the app's store page (if that's even a thing any more) to get it fully back to where it was.