App now respects phone color theme
Apparently the RequestedTheme setting (which is set by default!) will override the chosen color scheme (light/dark) from the phone's settings, which is kind of jacked up. Docs say there should be a "Default" value for the setting, but only Light or Dark seem to work. I'm not sure how to request a system-wide dark theme on Win10 for testing (or if you even can), but this works on the phone emulators and defaults to light on PC.