Main app screen now looks correct on all devices
I completely reformatted how the app was built in order to make it look and fit correctly on all different window sizes and devices. The old app was built by moving around all the images in the visual editor which set margins, widths, and heights automatically. Unfortunately that meant the app did not scale under the UWP model and ended up looking terrible. So I reworked everything with ViewBoxes (to maintain aspect ratio) and Grids (for layout) so that I could dump all the old method.

I also used some sweet content-aware-delete stuff in Gimp to make the navigation cross (up/down/left/right/ok) not have arrows already on it that needed to be lined up exactly with the actual button. And removed some extra space around the edge of the 'ok' button so it was more easily placed to cover up the built-in one (which I could have content-aware-deleted, but it was not looking amazing when I was trying it since it's so big).