Disabled frame counter
This isn't a game, so we're not really concerned about framerate (I'm not doing anything that would ever force the screen to update more than it has to). Enabling this was showing the framerate on a ton of different windows while the app was running and it was getting in the way of seeing meaningful information.