Reworked UI to not be dependent on ListBox
ListBox created odd side effects in UWP world like selecting a line of text and having the background highlight, etc. I'm not sure if I was overriding this style in the previous app or what, but it doesn't matter because the same interface/UX is just as easily implemented with ScrollViewer and StackPanel.