Fixed TV key controls flickering when first activated
Apparently assigning from the BitmapImage to the actual displayed Image is enough to cause a load and cache of the image. Doing this fixes the layout flickering when pressing an image for the first time (because the control momentarily had no width/height and the controls around it adjusted theirs to compensate) and the fact that the first press of a button didn't actually trigger the command (presumably because the pointer left the control's area when it went to size 0,0 which caused a cancel event instead of a tap).