Added support for displaying a YesNo box
In order for this to work the way I wanted it to (static method), there was a somewhat hacky requirement of storing a reference to the MainPage when it gets created. I think this will be okay because there's only ever one main page in the app, but it still feels a little gross. The only reason we need this is to get at a dispatcher when some non-UI thread wants to display a dialog box.