Don't crash when connection closes after pressing key
If the connection is closed for any reason between us pressing a key and trying to read its response, we need to ensure we handle the situation by notifying the user that we've lost connection to the device.

I also cleaned up a few cases that were only catching specific exceptions that I don't believe are possible to be thrown any more; for this app's case, we only care that the call succeeded or not, not what specific exception was thrown.