Visual overhaul to use SplitView
This modernizes the look of the app to use the left-side slide-out panel method that most modern apps use. Doing this required Page-ifying each section of controls so they could be navigated to in the central Frame, allowing the "go back" navigation style to work.

I also discovered that the ugly limitations of {Binding} could be mitigated somewhat with the use of {x:Bind} which allows compile-time binding checks (as opposed to {Binding}'s runtime-only checks) so I converted a few things over and was able to delete the XamlHelper glue code in the process.