Added option to always show the navbar
The navbar can now be either hidden and invoked only by the menu button, present in compact mode and expanded by the menu button, or always present and expanded, depending on settings and window width. Doing this properly meant getting a separate library to have "and" conditions on visual state triggers since the only thing Microsoft supports is "or" conditions.