Made NavMenu scrollable
On low-res devices turned to landscape mode, the nav items were either running into each other or inaccessible. This allows them to scroll so the user can reach them. I toyed with the idea of scaling them down so they continue to fit on the screen, but they looked pretty terrible scaled.