Fixed disconnect when sending a second key
Every time a key was sent, a new DataReader was being created on the socket with a 'using' statement. When this left scope, the 'using' would cause a Dispose of the DataReader which apparently takes the underlying stream with it. Embarrassing that this wasn't found until now. Bad testing on my part.